refactor(database): extract migration runner from connection.go
Extract migration-related methods into connection_migrations.go (353 lines):
Migrate, runUserMigrations, scanMigrationFiles, getAppliedMigrations,
applyFilesystemMigration, logMigrationExecution, grantRolesToRuntimeUser,
validateMigrationSQL, migrationFile type.

Connection.go reduced from 1076 to 735 lines.
